aligned team contributors partners card margins
