+++ src/ignitus-About/Styles/style.scss
@@ -3,7 +3,7 @@
 .about-wrapper {
 	font-size: $ignite-font-size-sm;
 	font-weight: $ignite-font-weight-normal;
-	background: url('/images/split-blue-background.png') no-repeat;
+	background: url('https://storage.googleapis.com/ignitus/Assets/split-blue-background.png') no-repeat;
 	background-size: 70%;
 	background-position: top right;
 }
