+++ src/ignitus-Contributors/Styles/style.scss
@@ -45,3 +45,7 @@
   border-width: 0px !important;
 }
 
+.card_row {
+    margin-right: 88px !important;
+    margin-left: 88px !important;
+}
