+++ src/ignitus-Introduction/Components/Constants.js
@@ -1,0 +1,1 @@
+export const resume 		= 'https://storage.googleapis.com/ignitus/Assets/resume.png'
