+++ src/ignitus-Partners/Styles/style.scss
@@ -47,3 +47,8 @@
   border-radius: 16px !important;
   border-width: 0px !important;
 }
+
+.card_row {
+    margin-right: 88px !important;
+    margin-left: 88px !important;
+}
