+++ src/ignitus-ProfessorLogin/Components/Constants.js
@@ -1,1 +1,4 @@
-export const logo 		= 'https://storage.googleapis.com/ignitus/IgnitusLogos/Black_logo.png'
+export const logo 					= 'https://storage.googleapis.com/ignitus/IgnitusLogos/Black_logo.png'
+export const professor_login 		= 'https://storage.googleapis.com/ignitus/Assets/professors-blue.png'
+
+
