+++ src/ignitus-ProfessorSignUp/Components/Constants.js
@@ -1,1 +1,2 @@
-export const logo 		= 'https://storage.googleapis.com/ignitus/IgnitusLogos/Black_logo.png'
+export const logo 					= 'https://storage.googleapis.com/ignitus/IgnitusLogos/Black_logo.png'
+export const professor_signup 		= 'https://storage.googleapis.com/ignitus/Assets/professor-signup.png'
