+++ src/ignitus-StudentLogin/Components/Constants.js
@@ -1,1 +1,2 @@
-export const logo 		= 'https://storage.googleapis.com/ignitus/IgnitusLogos/Black_logo.png'
+export const logo 				= 'https://storage.googleapis.com/ignitus/IgnitusLogos/Black_logo.png'
+export const student_login 		= 'https://storage.googleapis.com/ignitus/Assets/students-blue.png'
