+++ src/ignitus-StudentSignUp/Components/Constants.js
@@ -1,1 +1,3 @@
-export const logo 		= 'https://storage.googleapis.com/ignitus/IgnitusLogos/Black_logo.png'
+export const logo 					= 'https://storage.googleapis.com/ignitus/IgnitusLogos/Black_logo.png'
+export const student_signup 		= 'https://storage.googleapis.com/ignitus/Assets/student-signup.png'
+
