+++ src/ignitus-Team/Components/Constants.js
@@ -1,37 +1,18 @@
-const AVATARS = [{
-  title: 'Grant P. Her',
-  path: '/images/about/avatar-grant.png',
-  position: 'Product Manager',
-},
-{
-  title: 'Melanie Caruso',
-  position: 'Marketing Specialist',
-  path: '/images/about/avatar-melanie.png',
-},
-{
-  title: 'Nelson Bailey',
-  position: 'Founder',
-  path: '/images/about/avatar-nelson.png',
-}, {
-  title: 'Bertha Gowans',
-  position: 'Back-end Engineer',
-  path: '/images/about/avatar-bertha.png',
-}, {
-  title: 'George Bourgeois',
-  position: 'UI Designer',
-  path: '/images/about/avatar-george.png',
-}, {
-  title: 'Robert A. Pennell',
-  position: 'Data Scientist',
-  path: '/images/about/avatar-robert.png',
-}, {
-  title: 'Eugene Goebel',
-  position: 'ML Specialist',
-  path: '/images/about/avatar-eugene.png',
-}, {
-  title: 'Lauren Haskins',
-  position: 'UX Researcher',
-  path: '/images/about/avatar-lauren.png',
-}];
+export const Afelio    	= 'https://storage.googleapis.com/ignitus/Team/afelio.jpeg'
+export const Divyanshu  = 'https://storage.googleapis.com/ignitus/Team/divyanshu.jpeg'
+export const Ayush    	= 'https://storage.googleapis.com/ignitus/Team/ayush.jpeg'
+export const Jennifer   = 'https://storage.googleapis.com/ignitus/Team/jennifer.jpeg'
+export const Paarmita   = 'https://storage.googleapis.com/ignitus/Team/paarmita.jpeg'
+export const Pranjal 	= 'https://storage.googleapis.com/ignitus/Team/pranjal.jpeg'
+export const Hafsa    	= 'https://storage.googleapis.com/ignitus/Team/Hafsa.jpeg'
+export const Desire  	= 'hhttps://storage.googleapis.com/ignitus/Team/Desire.png'
 
-export default AVATARS;
+
+export const Avatar1    = 'https://storage.googleapis.com/ignitus/Team/avatar-bertha.png'
+export const Avatar2    = 'https://storage.googleapis.com/ignitus/Team/avatar-eugene.png'
+export const Avatar3    = 'https://storage.googleapis.com/ignitus/Team/avatar-george.png'
+export const Avatar4    = 'https://storage.googleapis.com/ignitus/Team/avatar-grant.png'
+export const Avatar5    = 'https://storage.googleapis.com/ignitus/Team/avatar-lauren.png'
+export const Avatar6    = 'https://storage.googleapis.com/ignitus/Team/avatar-melanie.png'
+export const Avatar7    = 'https://storage.googleapis.com/ignitus/Team/avatar-nelson.png'
+export const Avatar8    = 'https://storage.googleapis.com/ignitus/Team/avatar-robert.png'
