+++ src/ignitus-Team/Components/Data.js
@@ -1,0 +1,37 @@
+const AVATARS = [{
+  title: 'Grant P. Her',
+  path: '/images/about/avatar-grant.png',
+  position: 'Product Manager',
+},
+{
+  title: 'Melanie Caruso',
+  position: 'Marketing Specialist',
+  path: '/images/about/avatar-melanie.png',
+},
+{
+  title: 'Nelson Bailey',
+  position: 'Founder',
+  path: '/images/about/avatar-nelson.png',
+}, {
+  title: 'Bertha Gowans',
+  position: 'Back-end Engineer',
+  path: '/images/about/avatar-bertha.png',
+}, {
+  title: 'George Bourgeois',
+  position: 'UI Designer',
+  path: '/images/about/avatar-george.png',
+}, {
+  title: 'Robert A. Pennell',
+  position: 'Data Scientist',
+  path: '/images/about/avatar-robert.png',
+}, {
+  title: 'Eugene Goebel',
+  position: 'ML Specialist',
+  path: '/images/about/avatar-eugene.png',
+}, {
+  title: 'Lauren Haskins',
+  position: 'UX Researcher',
+  path: '/images/about/avatar-lauren.png',
+}];
+
+export default AVATARS;
