+++ src/ignitus-Team/Styles/style.scss
@@ -17,3 +17,8 @@
   border-radius: 16px;
   min-height: 14rem;
 }
+
+.team_row {
+	margin-right: 25px !important;
+    margin-left: 25px !important;
+}
